Fixed uniform attribution mistake on meanShiftShader.meanshift() and some texelFetching details on meanshift.frag
